Adding some abstract classes to remove duplicate code in the future and clean things up
