+++ build.py
@@ -5,5 +5,7 @@ PyInstaller.__main__.run([
     '--onefile',
     '--windowed',
     '--hidden-import=Import.Florida',
-    '--hidden-import=Warehouse.Florida'
+    '--hidden-import=Warehouse.Florida',
+    '--hidden-import=Import.State',
+    '--hidden-import=Warehouse.State'
 ])
